service/requirements.txt
# E-Banking Service Layer - Python Dependencies
# Install with: pip install -r requirements.txt

# FastAPI Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic[email]

# Environment variables
python-dotenv>=1.0.0

# Date/Time handling
python-dateutil>=2.8.0

# Testing (optional)
pytest>=7.0.0
httpx>=0.24.0
